moved all example tests in genbank package to example_test.go.
